More structure.
