--- Monopoly/src/monopoly/MonopolyRunner.java
@@ -1,21 +1,0 @@
-package monopoly;
-
-public class MonopolyRunner
-{
-	
-	public static void main(String[] args)
-	{
-		introToGame();
-		rulesOfTheGame();
-		
-		
-	}
-	
-	public static void introToGame(){
-//		Introduction.introToGame();
-	}
-	
-	public static void rulesOfTheGame() {
-//		Introduction.rules();
-	}
-}
