+++ Monopoly/src/objects/Go_To_Jail.java
@@ -1,6 +1,4 @@
-package monopoly;
-
-import objects.Player;
+package objects;
 
 public class Go_To_Jail extends Space{
