+++ Monopoly/src/objects/Luck.java
@@ -1,6 +1,4 @@
-package monopoly;
-
-import objects.Player;
+package objects;
 
 public class Luck extends Space{
