+++ Monopoly/src/program/Intitialize.java
@@ -1,0 +1,18 @@
+package program;
+
+import java.util.ArrayList;
+import objects.Player;
+import objects.Space;
+
+public class Intitialize
+{
+	int turn = 0;
+	static int numbOfPlayers;
+	static Space[] board;
+	static ArrayList<Player> players;
+	
+	public static void intitialize()
+	{
+		
+	}
+}
