+++ Monopoly/src/program/MonopolyRunner.java
@@ -5,9 +5,10 @@ public class MonopolyRunner
 	
 	public static void main(String[] args)
 	{
-		Intro.introToGame(); //ask to play
+		Intro.intro(); //ask to play
 		Settup.settup(); //ask numb of players and make players
-		Start.start(); //welcome to game and display rules if wanted
+		Start.start(); //welcome to game
+		Start.rules(); //display rules if wanted
 		Initialize.initialize(); //set up static variables and read property names
 		while(players.size() > 1)
 		{
